requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.2
flask-cors==5.0.1
Flask-SQLAlchemy==3.0.5
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==3.0.2
pillow==11.2.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==2.3.6

